+++ src/Stormancer.Plugins/Gog/Stormancer.Server.Plugins.Galaxy/Changelog.rst
@@ -14,7 +14,7 @@ Unreleased
 Changed
 *******
 - Specify CurrentPlatform when creating an user.
-- Update dependency to Stormancer.Abstraction.Server 7.0.0
+- Update dependency to Stormancer.Abstraction.Server 7.1.0
 
 2.0.2.1
 -------