+++ src/Stormancer.Plugins/Profile/Stormancer.Server.Plugins.Profile/Changelog.rst
@@ -13,7 +13,8 @@ Unreleased
 ----------
 Changed
 *******
-- Update dependency to Stormancer.Abstraction.Server 7.0.0
+- Update dependency to Stormancer.Abstraction.Server 7.1.0
+- Changed Profile dto to include a field stating if the profile was found or not
 Fixed
 *****
 - Create empty entries in GetProfiles for missing profiles.